requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
Babel>=1.3
eventlet>=0.15.1
keystonemiddleware>=1.0.0
oslo.config>=1.4.0.0a3
oslo.messaging>=1.4.0.0a3
